skills/repo-maintainer/requirements.txt
# Repository Maintainer Skill Dependencies
PyYAML>=6.0
